refactor(api): move password-generator service in IAM context
