+++ api/src/identity-access-management/domain/services/password-generator.service.js
